+++ AreaApi/package.json
@@ -10,9 +10,11 @@
   "dependencies": {
     "@octokit/rest": "^16.43.1",
     "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "discord.js": "^11.5.1",
     "env-cmd": "^10.0.1",
     "express": "^4.17.1",
+    "express-swagger-generator": "^1.1.17",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.5.2",
     "mongoose": "^5.8.10",
